* Fixed a compiler warning in sdk/editormanager.cpp. * When a file is saved, it is stripped of excessive spaces and unifies newlines used. Wrapped it all in a single undo action. * Added support for <Notice> nodes in templates, used for messages displayed when a new project is created by a template. Can be a child of <Template> or <Option>. * Added template files for Irrlicht (improved template originally created by Cmoibenlepro - patch #1275141).
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@832 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/editormanager.cpp
@@ -545,7 +545,7 @@ EditorBase* EditorManager::GetActiveEditor()
 void EditorManager::ActivateNext()
 {
     int sel = m_pNotebook->GetSelection();
-    if (sel < m_pNotebook->GetPageCount() - 1)
+    if (sel < (int)m_pNotebook->GetPageCount() - 1)
         ++sel;
     else
         sel = 0;
@@ -747,7 +747,7 @@ bool EditorManager::QueryClose(EditorBase *ed)
 
 int EditorManager::FindPageFromEditor(EditorBase* eb)
 {
-    for (int i = 0; i < m_pNotebook->GetPageCount(); ++i)
+    for (int i = 0; i < (int)m_pNotebook->GetPageCount(); ++i)
     {
         if (m_pNotebook->GetPage(i) == eb)
             return i;
